requirements.txt
streamlit>=1.32.0
chromadb>=0.5.0
scikit-learn>=1.3.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
google-genai>=1.0.0
